+++ packages/eslint-config-typescript/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+	extends: [
+		'@deloitte-digital-au/eslint-config',
+	],
+	env: {
+		node: true,
+	},
+};