+++ packages/eslint-config-typescript/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+	parser: '@typescript-eslint/parser',
+	extends: [
+		'@deloitte-digital-au/eslint-config',
+		'./rules/typescript',
+	].map(require.resolve),
+	plugins: ['@typescript-eslint'],
+};