+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### [UNPUBLISHED]
 
 - Enforce `async () => {}` instead of `async() => {}`.
+- Added `eslint-plugin-you-dont-need-lodash-underscore` plugin
+- Allow JSX in `.tsx` files
+- Added a TypeScript config
+- Migrated this mono repo from npm to yarn
 
 ### [3.4.0]